+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - This CHANGELOG file
+- New method ClientInterface::testCredentials()
 
 ## [0.1.0] - 2019-04-19